# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MTAA Fitness API",
        "description": "API for user registration, login, workouts, friends and plans.",
        "contact": {
            "email": "support@fitness-app.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/predefined-workouts": {
            "get": {
                "tags": [
                    "Predefined Workouts"
                ],
                "summary": "Get list of predefined workouts by level",
                "operationId": "ee16cf33fbb64bc60e0d69846a15a702",
                "parameters": [
                    {
                        "name": "level",
                        "in": "query",
                        "description": "Level of the workout (beginner or advanced)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "beginner",
                                "advanced"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of workouts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Full Body Beginner Workout"
                                            },
                                            "image": {
                                                "type": "string",
                                                "example": "http://localhost/images/predefined/workout.jpg"
                                            },
                                            "duration": {
                                                "type": "integer",
                                                "example": 30
                                            },
                                            "calories": {
                                                "type": "integer",
                                                "example": 200
                                            },
                                            "exercise_count": {
                                                "type": "integer",
                                                "example": 5
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid level"
                    }
                }
            }
        },
        "/api/predefined-workouts/{id}": {
            "get": {
                "tags": [
                    "Predefined Workouts"
                ],
                "summary": "Get detailed info about a predefined workout",
                "operationId": "ed6ef4fca548fa5fe2dc93b63a4e0bfd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the predefined workout",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Workout details with exercises",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Full Body Beginner Workout"
                                        },
                                        "focus": {
                                            "type": "string",
                                            "example": "Full Body"
                                        },
                                        "duration": {
                                            "type": "integer",
                                            "example": 30
                                        },
                                        "calories": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "sets_reps": {
                                            "type": "string",
                                            "example": "3 sets of 10 reps"
                                        },
                                        "rest": {
                                            "type": "string",
                                            "example": "30 seconds"
                                        },
                                        "benefits": {
                                            "type": "string",
                                            "example": "Improves endurance and strength"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "http://localhost/images/predefined/workout.jpg"
                                        },
                                        "exercises": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Push-Ups"
                                                    },
                                                    "reps_sets": {
                                                        "type": "string",
                                                        "example": "3 x 10"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Keep your back straight..."
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "http://localhost/images/predefined/exercises/pushups.jpg"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Workout not found"
                    }
                }
            }
        },
...
    },
...
}